Pass componments to separate files
